refactor: remove host_support observability module
