+++ host_support/__init__.py
@@ -1,12 +1,10 @@
 """Host-side shared helpers."""
 
 from .confirmation import is_confirmation_text
-from .observability import build_trace
 from .provider_mode import ProviderConfig, print_startup_config, resolve_provider_config
 
 __all__ = [
     "ProviderConfig",
-    "build_trace",
     "is_confirmation_text",
     "print_startup_config",
     "resolve_provider_config",
